+++ src/audio_metadata/__init__.py
@@ -3,7 +3,6 @@ from .api import *
 from .exceptions import *
 from .formats import *
 from .models import *
-from .warnings import *
 
 __all__ = [
 	*__about__.__all__,
@@ -11,5 +10,4 @@ __all__ = [
 	*exceptions.__all__,
 	*formats.__all__,
 	*models.__all__,
-	*warnings.__all__,
 ]