+++ src/audio_metadata/exceptions.py
@@ -1,42 +1,59 @@
 __all__ = [
 	'AudioMetadataException',
-	'InvalidBlock',
-	'InvalidChunk',
-	'InvalidComment',
-	'InvalidFormat',
-	'InvalidFrame',
-	'InvalidHeader',
+	'AudioMetadataWarning',
+	'FormatError',
+	'TagError',
 	'UnsupportedFormat',
 ]
 
+import sys
+import warnings
+
+
+##############
+# Exceptions #
+##############
 
 class AudioMetadataException(Exception):
-	"""Base exception class."""
-
-
-class InvalidBlock(AudioMetadataException):
-	"""Exception raised when a FLAC metadata block is invalid."""
-
-
-class InvalidComment(AudioMetadataException):
-	"""Exception raised when a Vorbis comment is invalid."""
-
-
-class InvalidChunk(AudioMetadataException):
-	"""Exception raised when a WAV chunk is invalid."""
-
-
-class InvalidFormat(AudioMetadataException):
-	"""Exception raised when a file format is invalid."""
-
-
-class InvalidFrame(AudioMetadataException):
-	"""Exception raised when an ID3v2 frame is invalid."""
-
-
-class InvalidHeader(AudioMetadataException):
-	"""Exception raised when a metadata header is invalid."""
+	"""Base exception for audio-metadata."""
 
 
 class UnsupportedFormat(AudioMetadataException):
-	"""Exception raised when loading a file that isn't supported."""
+	"""An unsupported format, version, or profile was encountered."""
+
+
+class FormatError(AudioMetadataException):
+	"""The binary format of a data input is invalid."""
+
+
+class TagError(AudioMetadataException):
+	"""A tag is not compliant to a specification."""
+
+
+############
+# Warnings #
+############
+
+# Override warning output format.
+def showwarning(message, category, filename, lineno, file=None, line=None):  # pragma: nocover
+	if file is None:
+		file = sys.stderr
+		if file is None:
+			return
+
+	delim = '\n    '
+	nl = '\n'
+	s = f"{category.__name__}:{delim}{delim.join(line for line in str(message).split(nl))}\n"
+
+	try:
+		file.write(s)
+	except (IOError, UnicodeError):
+		pass
+
+
+warnings.showwarning = showwarning
+del showwarning
+
+
+class AudioMetadataWarning(UserWarning):
+	"""Base warning for audio-metadata."""